removed connection pool references completely
