+++ AWSAppSyncClient/Subscription/Support/AppSyncLogger.swift
