+++ AWSAppSyncClient/Subscription/Support/RealtimeGatewayURLInterceptor.swift
