+++ AWSAppSyncClient/Subscription/Support/SubscriptionConnectionType.swift
