+++ AWSAppSyncClient/Subscription/Support/SubscriptionConstants.swift
